refactor: 기존 UI 관련 파일 위치 변경
